Minor bug fix to recent changes

+++ src/CanSettings.h
@@ -41,8 +41,8 @@ struct CanTiming
 
 	void SetDefaults_1Mb() noexcept
 	{
-		period = (uint32_t)((ClockFrequency + (bitRate/2))/bitRate);
-		tseg1 = (uint32_t)(period * DefaultSamplePoint) - 1;				// this excludes the 1-clock sync phase for historical reasons, hence the -1
+		period = (uint16_t)((ClockFrequency + (bitRate/2))/bitRate);
+		tseg1 = (uint16_t)(period * DefaultSamplePoint) - 1;				// this excludes the 1-clock sync phase for historical reasons, hence the -1
 		jumpWidth = period - (tseg1 + 1);									// this is the maximum possible, as recommended by CiA
 	}
 };
